+++ CHANGELOG.md
@@ -6,6 +6,13 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic
 Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Changed
+
+- Unsafe chunks have been separated out into the `sized-chunks` crate, which is
+  now a dependency of `im`.
+
 ## [12.3.0] - 2019-01-15
 
 ### Added